Fix UInt32 to Int32 conversion for Linux socket creation
- Cast SOCK_STREAM.rawValue to Int32 on Linux/Glibc systems
- Resolves type conversion error on Ubuntu builds
- Maintains Darwin compatibility with direct SOCK_STREAM usage

+++ Sources/TinyHttpServer/TinyHttpServer.swift
@@ -33,7 +33,7 @@ public class TinyHttpServer {
         #if canImport(Darwin)
         let serverSocket = socket(AF_INET, SOCK_STREAM, 0)
         #elseif canImport(Glibc)
-        let serverSocket = socket(AF_INET, SOCK_STREAM.rawValue, 0)
+        let serverSocket = socket(AF_INET, Int32(SOCK_STREAM.rawValue), 0)
         #endif
         guard serverSocket != -1 else { throw TinyHttpServerError.socketCreationFailed }
